requirements/base.txt
django==4.2.1
djangorestframework==3.14.0
asgiref==3.7.2
psycopg2-binary==2.9.6
sqlparse==0.4.4
asgiref==3.7.2



gunicorn==20.1.0
pytz==2023.3
typing_extensions==4.5.0
flake8==6.0.0
Pillow==9.5.0
redis==4.5.5